[Decomp][BattleGround] Cohesion split: extract rewards to BattleGroundReward.cpp
